requirements.txt
python-telegram-bot
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv